api/requirements.txt
python-pptx==0.6.23
anthropic==0.18.1
